awnsystemmonitor:       	-fix for https://bugs.launchpad.net/awn-extras/+bug/221345 (needs someone with lots of memory to test.  need >2 Gig process)

+++ src/awnsystemmonitor/awntop_cairo_component.c
@@ -95,8 +95,8 @@ typedef struct
     int     uid;
     int     pri;
     int     nice;
-    guint    virt;
-    guint    res;
+    guint64    virt;
+    guint64    res;
     long    shr;
     long     cpu;
     guint64     mem;
